Generalized Order-Restricted Information Criterion (GORIC) value for a set of hypotheses in multivariate regression models

Computes the Generalized Order-Restricted Information Criterion (GORIC) values for a set of hypotheses in multivariate regression models. The GORIC is a modification of the AIC (Akaike, 1973) and the ORIC (Anraku, 1999) such that it can be applied to a general form of order-restricted hypotheses in multivariate normal linear models (Kuiper, Hoijtink, and Silvapulle, 2011; Kuiper, Hoijtink, and Silvapulle, unpublished).

Details

Package: 	 goric

Type: 	 Package

Version: 	 0.0-8

Date: 	 2013-03-26

License: 	 GPL (>=2)

LazyLoad: 	 yes

In a \(t\)-variate regression with \(k\) predictors (including an optional intercept), the order-restricted hypotheses should be of the form: \(Constr \theta \geq rhs\), with \(\theta\) a vector of length \(tk\), where the first \(k\) elements belong to first dependent variable, ..., and the last \(k\) elements to the last dependent variable; \(rhs\) is a vector of length \(c\); and \(Constr\) a \(c \times tk\) matrix of full rank.

Thera are two requirements:

The first \(nec\) constraints must be the equality contraints (i.e., \(Constr[1:nec, 1:tk] \theta = rhs[1:nec]\)) and the remaing ones the inequality contraints (i.e., \(Constr[nec+1:c, 1:tk] \theta \geq rhs[nec+1:c]\)).

When \(rhs\) is not zero, \(Constr\) should be of full rank (after discarding redundant restrictions).

More information can be obtained from Kuiper, Hoijtink, and Silvapulle (2011), Kuiper, Hoijtink, and Silvapulle (2012), and Kuiper and Hoijtink (unpublished).

orlm renders the order-restricted maximum likelihood estimates (i.e., the maximum likelihood estimates subject to the hypothesis of interest) and the corresponding log likelihood for the hypothesis of interest (defined by \(Constr\), \(rhs\), and \(nec\)). Additionally it gives the (unconstrained) maximum likelihood estimates and the active contraints. goric gives the order-restricted log likelihood, the penalty of the GORIC, the GORIC values, and the GORIC weights for a set of hypotheses (orlm objects), where the penalty is based on \(iter\) iterations.

The hypothesis with the lowest GORIC value is the preferred one.

The GORIC weights reflect the support of each hypothesis in the set. To compare two hypotheses (and not one to the whole set), one should examine the ratio of the two corresponding GORIC weights.

To safequard for weak hypotheses (i.e., hypotheses not supported by teh data), one should include a model with no constraints (the so-called unconstrained model). More information can be obtained from Kuiper, Hoijtink, and Silvapulle (2011) and Kuiper, Hoijtink, and Silvapulle (unpublished).

References

Akaike, H. (1973). Information theory and an extension of the maximum likelihood principle. In Proc. 2nd Int. Symp. Information Theory, Ed. B. N. Petrov and F. Csaki, pp. 267/226-81. Budapest: Akademiai kiado.

Anraku K. (1999). An Information Criterion for Parameters under a Simple Order Restriction. Biometrika, 86, 141--152.

Kuiper R.M., Hoijtink H., Silvapulle M.J. (2011). An Akaike-type Information Criterion for Model Selection Under Inequality Constraints. Biometrika, 98, 495--501.

Kuiper R.M., Hoijtink H., Silvapulle M.J. (2012). Generalization of the Order-Restricted Information Criterion for Multivariate Normal Linear Models. Journal of Statistical Planning and Inference, 142, 2454-2463. doi:10.1016/j.jspi.2012.03.007.

Kuiper R.M. and Hoijtink H. (submitted to Journal of Statictical Software). A Fortran 90 Program for the Generalization of the Order-Restricted Information Criterion.