GORIC penalty term

Calculates the GORIC penalty term (level probability) by Monte-Carlo simulation.

goric_penalty(object, iter = 1e+05, type = "GORIC", mc.cores = 1)

orglm_penalty(object, iter = 1e+05, type = "GORIC", mc.cores = 1)

Arguments

object
an object of class orlm, orgls (or orglm for function orglm_penalty)

iter
number of iterations to calculate GORIC penalty terms

type
if "GORIC" (default), the penalty term for the generalized order restriction information criterion is computed; with "GORICCa" or "GORICCb" small sample corrections for the penalty term are applied

mc.cores
number of cores using a socket cluster implemented in package parallel

See also

orlm, orgls, orglm