Fitting generalized least squares regression models with order restrictions

It determines the order-restricted maximum likelihood estimates and the corresponding log likelihood for the hypothesis of interest. Additionally it gives the (unconstrained) maximum likelihood estimates and the active contraints.

orglsSet(formula, data, weights=NULL,
         correlation=NULL, set, direction="increase",
         n=NULL, base=1, control = orlmcontrol())

Arguments

formula
an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.

data
an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.

weights
a varClasses object; more details are provided on the help pages in R package nlme

correlation
a corClasses object; more details are provided on the help pages in R package nlme

set
either a character string (see constrSet), or a list with slots for constr, rhs, and nec similarly defined as in orlm

direction
direction of the order constraints

n
a (possibly named) vector of sample sizes for each group

base
column of the constraint matrix representing a control group

control
a list of control arguments; see orlmcontrol for details.

Details

This function is just a wrapper for repeated calls of orgls with different constraint definitions. Predefined lists with constraint-sets can be constructed with function constrSet.

Value

an object of class list

References

Kuiper R.M., Hoijtink H., Silvapulle M.J. (2011). An Akaike-type Information Criterion for Model Selection Under Inequality Constraints. Biometrika, 98, 495--501.

Kuiper R.M., Hoijtink H., Silvapulle M.J. (2012). Generalization of the Order-Restricted Information Criterion for Multivariate Normal Linear Models. Journal of Statistical Planning and Inference, 142, 2454-2463. doi:10.1016/j.jspi.2012.03.007.

Kuiper R.M. and Hoijtink H. (submitted). A Fortran 90 Program for the Generalization of the Order-Restricted Information Criterion. Journal of Statictical Software.

See also

orlm, constrSet, goric